avcodec/h264_ps: fix storage size for offset_for_ref_frame
The spec defines the valid range of values to be INT32_MIN + 1 to INT32_MAX, inclusive.

+++ libavcodec/h264_ps.h
@@ -81,7 +81,7 @@ typedef struct SPS {
     uint32_t num_units_in_tick;
     uint32_t time_scale;
     int fixed_frame_rate_flag;
-    short offset_for_ref_frame[256]; // FIXME dyn aloc?
+    int32_t offset_for_ref_frame[256];
     int bitstream_restriction_flag;
     int num_reorder_frames;
     int scaling_matrix_present;
